add: MyPageContainer 생성
