+++ src/pages/Home.jsx
@@ -1,17 +1,26 @@
 import { useState, useEffect } from "react";
-import SHINE_DATA from "../mock";
+// import SHINE_DATA from "../mock";
 import MainPageInput from "../components/MainPageInput";
 import MainPagePosts from "../components/MainPagePosts";
 import supabase from "../supabaseClient";
 
 const Home = () => {
-  const [posts, setPosts] = useState(() => {
-    const selectPost = async () => {
-      const { data, error } = await supabase.from("posts").select("*");
+  const [posts, setPosts] = useState([]);
 
-      console.log("data", data);
+  // 포스팅 한  DB (이미지 X)
+  useEffect(() => {
+    const fetchPosts = async () => {
+      const { data, error } = await supabase.from("posts").select("*").order("created_at", { ascending: false });
+      if (error) {
+        console.error("Error fetching posts:", error.message);
+      } else {
+        setPosts(data);
+      }
     };
-  });
+
+    fetchPosts();
+  }, []);
+
   const addPosthandler = (data) => {
     setPosts([data, ...posts]);
   };
