+++ src/supabaseClient.js
@@ -1,5 +1,7 @@
 import { createClient } from "@supabase/supabase-js";
 
+console.log("import.meta.env.VITE_SUPABASE_URL", import.meta.env.VITE_SUPABASE_URL);
+
 const SUPABASE_URL = import.meta.env.VITE_SUPABASE_URL;
 const SUPABASE_KEY = import.meta.env.VITE_SUPABASE_KEY;
